Refactoring FileExtractor class for more modularity
