+++ cli/src/main/java/com/devonfw/tools/ide/commandlet/FileExtractor/ExctractorFileType.java
@@ -1,0 +1,5 @@
+package com.devonfw.tools.ide.commandlet.FileExtractor;
+
+enum ExtractorFileType {
+  ZIP, JAR, DMG, MSI, PKG
+}
